Don't draw any graphs if the data sources are not set. Turn the formal dataSource and delegate protocols into informal protocols.

+++ Source/userinterface/cocoa/3rd party/GraphX/Source/DataSources.m
@@ -9,7 +9,7 @@
 
 
 //************Curve DataSource
-- (float)yValueForXValue:(double)x
+- (double)yValueForXValue:(double)x
 {
 //  if( (x > -2 && x < -1) || (x > 1 && x < 2) )
 //  return INFINITY;
